entrance-microservice/requirements.txt
asgiref==3.5.1
auth-lib-services==0.0.2
certifi==2021.10.8
charset-normalizer==2.0.12
colorama==0.4.4
Django==4.0.4
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
idna==3.3
Pillow==9.1.0
PyJWT==1.7.1
python-dotenv==0.20.0
pytz==2022.1
qrcode==7.3.1
requests==2.27.1
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
